% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ovarian.R
\docType{data}
\name{Agilent.dat}
\alias{Agilent.dat}
\title{Ovarian cancer data measured from Agilent micro-array platform}
\format{
A data frame with 11861 rows and 306 columns:
\describe{
  The expression matrix with the gene in rows and samples in column
}
}
\source{
Provided by Mayo clinic
}
\usage{
Agilent.dat
}
\description{
The matrix of gene expression of samples
}
\keyword{datasets}
