% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ovarian.R
\docType{data}
\name{Agilent.recur}
\alias{Agilent.recur}
\title{Ovarian cancer data measured from Agilent micro-array platform}
\format{
A data frame with 306 rows and 1 variables:
\describe{
  \item{\code{recur}}{character Whether the samples from Agilent platform are recurrence cases}
}
}
\source{
Provided by Mayo clinic
}
\usage{
Agilent.recur
}
\description{
The recurrence status of samples: yes or no
}
\keyword{datasets}
