% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ovarian.R
\docType{data}
\name{Agilent.subtype}
\alias{Agilent.subtype}
\title{Ovarian cancer data measured from Agilent micro-array platform}
\format{
A data frame with 306 rows and 1 variables:
\describe{
  \item{\code{ordered.subtype}}{character Cancer subtype variables}
}
}
\source{
Provided by Mayo clinic
}
\usage{
Agilent.subtype
}
\description{
The cancer subtype variables: C1-MES, C2-IMM, C4-DIF, C5-PRO
}
\keyword{datasets}
