% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorFunc.R
\name{batch.correct.r1}
\alias{batch.correct.r1}
\title{Generic maximum likelihood estimation. The objective is the same with remeasured method
but the generic optimization methods are used like BFGS or conjugate gradient (CG)}
\usage{
batch.correct.r1(Y, X, Z, ind.r, Y.r, start = NULL, method = "BFGS")
}
\arguments{
\item{Y}{the response vector}

\item{X}{binary vector indicate the control or treatment}

\item{Z}{covariate matrix}

\item{ind.r}{index of remeasured samples}

\item{Y.r}{the response vector of remeasured sample}

\item{start}{the initial value of parameters, default is NULL}

\item{method}{the optimization method chosen for the optimization, default is BFGS}
}
\value{
The estimated coefficient of MLE
}
\description{
Generic maximum likelihood estimation. The objective is the same with remeasured method
but the generic optimization methods are used like BFGS or conjugate gradient (CG)
}
