% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IgnoreBatch.R
\name{batch.Ignore.S1}
\alias{batch.Ignore.S1}
\title{Ignore the batch effect}
\usage{
batch.Ignore.S1(Y, X, Z)
}
\arguments{
\item{Y}{the response vector}

\item{X}{binary vector indicate the control or treatment}

\item{Z}{covariate matrix}
}
\value{
The estimated coefficient
}
\description{
Ignore the batch effect
}
