% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReMeasure_S1.R
\name{batch.ReMeasure.S1}
\alias{batch.ReMeasure.S1}
\title{Our ReMeasure for batch effect elimination}
\usage{
batch.ReMeasure.S1(Y, X, Z, ind.r, Y.r)
}
\arguments{
\item{Y}{the response vector}

\item{X}{binary vector indicate the control or treatment}

\item{Z}{covariate matrix}

\item{ind.r}{index of remeasured samples}

\item{Y.r}{the response vector of remeasured sample}
}
\value{
The estimated coefficient of MLE
}
\description{
Our ReMeasure for batch effect elimination
}
\examples{
#' n = 100
n1 = 40
r1 = 1
r2 = 0.6
a0 = 0.25
a1 = 0.5

X =  as.numeric(gl(2, n / 2)) - 1
Z <- cbind(rep(1, n), rnorm(n))
b <- c(0, -0.5)
v1 = 2/(1+r1)
v2 = r1 * v1
Et <- rnorm(n, sd = ifelse (X == 0, sqrt(v1), sqrt(v2)))
Y <- Z \%*\% b + cbind(X, X) \%*\% c(a0, a1) + Et
Z.r.a <- Z[1 : (n / 2), ]
Et.r.a <- Et[1 : (n / 2)]
Y.r.a <- a1 + Z.r.a \%*\% b + r2 * sqrt(v2) * Et.r.a/ sqrt(v1) +
rnorm(n/2, sd = sqrt( (1 - r2^2) * v2 ) )
ind.r <- 1:n1
Y.r = Y.r.a[ind.r]

Estimate = batch.ReMeasure.S1(Y, X, Z, ind.r, Y.r)
Estimate$a0
Estimate$a1
Estimate$a0Var
Estimate$p.value
}
