% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{batch.ReMeasure.S1.res}
\alias{batch.ReMeasure.S1.res}
\title{Residual bootstrap method to estimate the uncertainty.}
\usage{
batch.ReMeasure.S1.res(Y, X, Z, ind.r, Y.r)
}
\arguments{
\item{Y}{the response vector}

\item{X}{binary vector indicate the control or treatment}

\item{Z}{covariate matrix}

\item{ind.r}{index of remeasured samples}

\item{Y.r}{the response vector of remeasured sample}
}
\value{
The estimated coefficient and ztest statistic given by ReMeasure method and
the residual boostrap statistic ztestb. One can use the statistic to
estimate the power
}
\description{
Residual bootstrap method to estimate the uncertainty.
}
