% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerCal.R
\name{Calculate_Power_S1}
\alias{Calculate_Power_S1}
\title{Power calculator}
\usage{
Calculate_Power_S1(nc1, nt2, r1, r2, a0, a1, nc2, seedJ = 2)
}
\arguments{
\item{nc1}{size of control sample in batch 1}

\item{nt2}{size of treatment sample in batch 2}

\item{r1}{scale effect that determines the variances}

\item{r2}{correlation parameter}

\item{a0}{true effect}

\item{a1}{batch effect}

\item{nc2}{remeasured control size of control sample}

\item{seedJ}{choose the seed}
}
\value{
The power calculated from formula
}
\description{
Power calculator
}
\examples{
alpha = 0.05
Calculate_Power_S1(nc1 = 50, nt2 = 50, r1=0.5, r2=0.6, a0 = 0.5, a1 = 0.5,
nc2 = 20)
}
