% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ovarian.R
\docType{data}
\name{remeasure.recur}
\alias{remeasure.recur}
\title{Common remeasured samples between the two data set}
\format{
A data frame with 47 rows and 1 variables:
\describe{
  \item{\code{recur}}{character Whether the samples are recurrence cases}
}
}
\source{
Provided by Mayo clinic
}
\usage{
remeasure.recur
}
\description{
The recurrence status of samples: yes or no
}
\keyword{datasets}
