% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ovarian.R
\docType{data}
\name{remeasure.subtype}
\alias{remeasure.subtype}
\title{Common remeasured samples between the two data set}
\format{
A data frame with 47 rows and 1 variables:
\describe{
  \item{\code{ordered.subtype}}{character Cancer subtype variables}
}
}
\source{
Provided by Mayo clinic
}
\usage{
remeasure.subtype
}
\description{
The cancer subtype variables: C1-MES, C2-IMM, C4-DIF, C5-PRO
}
\keyword{datasets}
