% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ovarian.R
\docType{data}
\name{RNAseq.dat}
\alias{RNAseq.dat}
\title{Ovarian cancer data measured from Agilent RNA-Seq platform}
\format{
A data frame with 11861 rows and 97 variables:
\describe{
  The expression matrix with the gene in rows and samples in column
}
}
\source{
Provided by Mayo clinic
}
\usage{
RNAseq.dat
}
\description{
The matrix of gene expression of samples
}
\keyword{datasets}
