% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ovarian.R
\docType{data}
\name{RNAseq.recur}
\alias{RNAseq.recur}
\title{Ovarian cancer data measured from Agilent RNA-Seq platform}
\format{
A data frame with 97 rows and 1 variables:
\describe{
  \item{\code{recur}}{character Whether the samples from RNASeq platform are recurrence cases}
}
}
\source{
Provided by Mayo clinic
}
\usage{
RNAseq.recur
}
\description{
The recurrence status of samples: yes or no
}
\keyword{datasets}
