% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch2Only.R
\name{batch.Batch2.S1}
\alias{batch.Batch2.S1}
\title{Only consider the samples in batch 2}
\usage{
batch.Batch2.S1(Y, X, Z, ind.r, Y.r)
}
\arguments{
\item{Y}{the response vector}

\item{X}{binary vector indicate the control or treatment}

\item{Z}{covariate matrix}

\item{ind.r}{index of remeasured sample points}

\item{Y.r}{the response vector of remeasured sample}
}
\value{
The estimation of of parameters include true parameter,
batch effect, variances and correlation
}
\description{
Only consider the samples in batch 2
}
